requirements.txt
# Affilify TikTok System - Python Dependencies
# Install with: pip3 install -r requirements.txt

# Core dependencies
python-dotenv==1.0.0

# Database
# (Using built-in sqlite3)

# Video processing
moviepy==1.0.3
Pillow==10.1.0

# Browser automation
playwright==1.40.0

# HTTP requests
requests==2.31.0

# AI/ML
google-generativeai==0.3.1

# Data processing
pandas==2.1.4
numpy==1.26.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
